Jul 15: [FIX] Bug Fixed 'advanced_loan_management'

+++ advanced_loan_management/models/loan_type.py
@@ -30,7 +30,7 @@ class LoanTypes(models.Model):
     _description = 'Loan Type'
 
     name = fields.Char(string='Name', help="LoanType Name")
-    loan_amount = fields.Integer(string='Loan Amount', help="Loan Amount")
+    loan_amount = fields.Float(string='Loan Amount', help="Loan Amount")
     tenure = fields.Integer(string='Tenure', default='1',
                             help="Amortization period")
     tenure_plan = fields.Char(string="Tenure Plan", default='monthly',
